error checking

+++ app.py
@@ -183,7 +183,13 @@ def removeFriend():
         query = "DELETE FROM Belong WHERE username = %s and groupName = %s "
         with connection.cursor() as cursor:
             cursor.execute(query, (username, group))
-
+        query = "SELECT Follow.followerUsername From Follow WHERE Follow.followerUsername = %s and Follow.followeeUsername = %s"
+        with connection.cursor() as cursor:
+            cursor.execute(query, (session["username"], username))
+        if (len(cursor.fetchall()) == 0):
+            query = "DELETE FROM Tag WHERE username = %s AND photoID in (select * from (SELECT photoID from tag NATURAL JOIN Photo where photoOwner = %s) as t)"
+            with connection.cursor() as cursor:
+                cursor.execute(query, (session["username"], username))
     return redirect("/friends")
 
 
@@ -274,8 +280,12 @@ def unfollow():
     with connection.cursor() as cursor:
         print(session['username'], username)
         cursor.execute(query, (session["username"], username))
-    query = "DELETE FROM Tag WHERE username = %s AND photoID in (select * from (SELECT photoID from tag NATURAL JOIN Photo where photoOwner = %s) as t)"
+    query = "SELECT groupName FROM Belong WHERE username = %s and groupName in (SELECT groupName FROM Belong WHERE username = %s) as t"
     with connection.cursor() as cursor:
+            cursor.execute(query, (session["username"], username))
+    if (len(cursor.fetchall()) == 0):
+        query = "DELETE FROM Tag WHERE username = %s AND photoID in (select * from (SELECT photoID from tag NATURAL JOIN Photo where photoOwner = %s) as t)"
+        with connection.cursor() as cursor:
             cursor.execute(query, (session["username"], username))
     return redirect("/follow")
